requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.117.1
flatbuffers==25.9.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.1
h11==0.16.0
h5py==3.14.0
idna==3.10
importlib_metadata==8.7.0
keras==3.10.0
libclang==18.1.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.1.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
protobuf==4.25.8
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
requests==2.32.5
rich==14.1.0
sniffio==1.3.1
starlette==0.48.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
Werkzeug==3.1.3
wrapt==1.17.3
zipp==3.23.0

